+++ app/assets/stylesheets/application.scss
@@ -16,5 +16,6 @@
 @import "layouts/*";
 @import "mixins/*";
 @import "modules/*";
+@import "product_owners/*";
 @import "products/*";
 @import "uswds-overrides/*";
