+++ app/assets/stylesheets/layouts/_application.scss
@@ -1,6 +1,27 @@
 .default-layout {
   background-color: $base-background-color;
 
+  .breadcrumbs {
+    font-size: 0.95em;
+    margin-bottom: 2em;
+
+    a {
+      border-bottom: 1px solid $medium-gray;
+      color: darken($medium-gray, 10);
+
+      &:hover {
+        border-bottom: 1px solid $color-primary;
+        color: $color-primary;
+        text-decoration: none;
+      }
+    }
+
+    i {
+      color: $medium-gray;
+      margin: 0 0.5em;
+    }
+  }
+
   .modal {
     @extend %modal;
   }
