+++ app/assets/stylesheets/modules/_navigation.scss
@@ -16,6 +16,7 @@
 
 .app-nav {
   background-color: $white;
+  box-shadow: 0 0 2px rgba($black, 0.15);
   padding: 0.9em 0;
 
   li {
