+++ app/assets/stylesheets/products/_index.scss
@@ -62,6 +62,7 @@
 
   .app-nav {
     background: none;
+    box-shadow: none;
   }
 
   .banner {
