+++ app/controllers/product_owners/dashboard_controller.rb
@@ -1,0 +1,7 @@
+module ProductOwners
+  class DashboardController < ProductOwnersController
+    def index
+      @product_requests = current_product_owner.product_requests
+    end
+  end
+end
