+++ app/controllers/product_owners/registrations_controller.rb
@@ -1,23 +1,21 @@
-module ProductOwners
-  class RegistrationsController < RegistrationsController
-    respond_to :html, :json
-    layout "minimal"
+class ProductOwners::RegistrationsController < RegistrationsController
+  respond_to :html, :json
+  layout "minimal"
 
-    private
+  private
 
-    def account_update_params
-      params.require(:product_owner).
-        permit(
-          :first_name,
-          :last_name,
-          :email,
-          :password,
-          :current_password)
-    end
+  def account_update_params
+    params.require(:product_owner).
+      permit(
+        :first_name,
+        :last_name,
+        :email,
+        :password,
+        :current_password)
+  end
 
-    def sign_up_params
-      params.require(:product_owner).
-        permit(:first_name, :last_name, :email, :password)
-    end
+  def sign_up_params
+    params.require(:product_owner).
+      permit(:first_name, :last_name, :email, :password)
   end
 end
