+++ app/models/dashboard.rb
@@ -1,0 +1,2 @@
+class Dashboard
+end
