+++ app/views/product_owners/dashboard/_product_request.html.haml
@@ -1,0 +1,4 @@
+= content_tag_for :tr, product_request do
+  %td
+    = link_to product_request.product.name,
+      product_path(product_request.product)
