+++ app/views/product_owners/dashboard/index.html.haml
@@ -1,0 +1,5 @@
+%section.product-requests.usa-grid
+  %header.page-header
+    %h1
+      = t(".heading")
+  = render "product_requests", product_requests: @product_requests
