+++ app/views/products/show.html.haml
@@ -1,5 +1,5 @@
 .usa-grid
-  .product-nav
+  .product-nav.breadcrumbs
     = link_to t(".products"), root_path
     %i.fa.fa-caret-right
     = @product.name
